bugfix: unify() needs to reset the index of the DataFrame
- unify() should result in two GraphFrames with the same structured index.
- implementation was not calling set_index() after reassinging the "node" column
- add set_index calls and add a test.
